CVS is visible, and reedirects, but I have permissions problems
