+++ bika/lims/browser/batchfolder.py
@@ -123,100 +123,6 @@ class BatchFolderContentsView(BikaListingView):
         return items
 
 
-class BatchFolderExportCSV(BatchFolderContentsView):
-
-    def __call__(self, REQUEST, RESPONSE):
-        """
-        Export invoice batch in csv format.
-        Writes the csv file into the RESPONSE to allow
-        the file to be streamed to the user.
-        Nothing gets returned.
-        """
-        import pdb;pdb.set_trace()
-        import csv
-        from cStringIO import StringIO
-        delimiter = ','
-        filename = 'batch.txt'
-
-        container = self.unrestrictedTraverse(REQUEST.get('current_path'))
-        assert container
-
-        container.plone_log("Exporting InvoiceBatch to CSV format for PASTEL")
-        invoices = container.listFolderContents()
-
-        if not len(invoices):
-            container.plone_log("InvoiceBatch contains no entries")
-
-        rows = []
-        _ordNum = 'starting at none'
-        for invoice in invoices:
-            new_invoice = True
-            _invNum = "%s" % invoice.getId()[:8]
-            _clientNum = "%s" % invoice.getClient().getAccountNumber()
-            _invDate = "%s" % invoice.getInvoiceDate().strftime('%Y-%m-%d')
-            _monthNum = invoice.getInvoiceDate().month()
-            if _monthNum < 7:
-                _periodNum = _monthNum + 6
-            else:
-                _periodNum = _monthNum - 6
-            _periodNum = "%s" % _monthNum
-        import pdb;pdb.set_trace()
-    #         _message1 = ''
-    #         _message2 = ''
-    #         _message3 = ''
-
-    # items = invoice.invoice_lineitems # objectValues('InvoiceLineItem')
-    #         mixed = [(item.getClientOrderNumber(), item) for item in items]
-    #         mixed.sort()
-    #         lines = [t[1] for t in mixed]
-
-    # iterate through each invoice line
-    #         for line in lines:
-    #             if new_invoice or line.getClientOrderNumber() != _ordNum:
-    #                 new_invoice = False
-    #                 _ordNum = line.getClientOrderNumber()
-
-    # create header csv entry as a list
-    #                 header = [ \
-    #                     "Header", _invNum, " ", " ", _clientNum, _periodNum, \
-    #                     _invDate, _ordNum, "N", 0, _message1, _message2, \
-    #                     _message3, "", "", "", "", "", "", 0, "", "", "", "", \
-    #                     0, "", "", "N"]
-    #                 rows.append(header)
-
-    #             _quant = 1
-    #             _unitp = line.getSubtotal()
-    #             _inclp = line.getTotal()
-    #             _item = line.getItemDescription()
-    #             _desc = "Analysis: %s" % _item[:40]
-    #             if _item.startswith('Water') or _item.startswith('water'):
-    #                 _icode = "002"
-    #             else:
-    #                 _icode = "001"
-    #             _ltype = "4"
-    #             _ccode = ""
-    # create detail csv entry as a list
-    #             detail = ["Detail", 0, _quant, _unitp, _inclp, \
-    #                       "", "01", "0", "0", _icode, _desc, \
-    #                       _ltype, _ccode, ""]
-    #             rows.append(detail)
-    # convert lists to csv string
-    #     ramdisk = StringIO()
-    #     writer = csv.writer(ramdisk, delimiter = delimiter)
-    #     assert(writer)
-    #     writer.writerows(rows)
-    #     result = ramdisk.getvalue()
-    #     ramdisk.close()
-    # stream file to browser
-    #     setheader = RESPONSE.setHeader
-    #     setheader('Content-Length', len(result))
-    #     setheader('Content-Type',
-    #         'text/x-comma-separated-values')
-    #     setheader('Content-Disposition', 'inline; filename=%s' % filename)
-    #     RESPONSE.write(result)
-    #     return
-
-
 class ajaxGetBatches(BrowserView):
 
     """ Vocabulary source for jquery combo dropdown box
