+++ bika/lims/browser/invoicebatch.zcml
@@ -18,4 +18,12 @@
       layer="bika.lims.interfaces.IBikaLIMS"
     />
 
+    <browser:page
+      for="bika.lims.interfaces.IInvoiceBatch"
+      name="export_batch_csv"
+      class=".invoicebatch.BatchFolderExportCSV"
+      permission="bika.lims.ManageInvoices"
+      layer="bika.lims.interfaces.IBikaLIMS"
+    />
+
 </configure>
