+++ bika/lims/upgrade/to318.py
@@ -16,7 +16,7 @@ def upgrade(tool):
     # Updated profile steps
     setup.runImportStepFromProfile('profile-bika.lims:default', 'jsregistry')
     setup.runImportStepFromProfile('profile-bika.lims:default', 'typeinfo')
-    setup.runImportStepFromProfile('profile-bika.lims:default', 'workflow_csv')
+    setup.runImportStepFromProfile('profile-bika.lims:default', 'workflow-csv')
     # Update workflow permissions
     wf = getToolByName(portal, 'portal_workflow')
     wf.updateRoleMappings()
